modules/barrier: [cleanup] handle ENOMEM in request
Problem: barrier creation exits on out of memory error.

Restructure barrier.enter handler so that an ENOMEM
error is returned to the user.

Also, move the hash insertion out of the barrier creation
function for clarity.

This is the last use of oom(), xzmalloc(), and xstrdup(),
so drop those includes.