libsubprocess: Remove invalid EOF calculation
In local subprocesses, it was assumed that an empty output buffer
for an exited subprocesses was equivalent to an EOF being generated
on that output stream.

This assumption was invalid.  The assumption was based on the traditional
logic that when you read() zero bytes on a file descriptor, you've
read EOF on a stream.  However, we are reading from a buffer and not a
stream.  While the subprocess has exited, it's possible that more data may
be available for the buffer later.  Since data and child exit status
go through a reactor, it is possible the subprocess exit status could be
seen before some final data is available in the buffer.

Fixes #2228