job-manager: include job metadata in hello response
Problem: flux-sched wants to reconstruct queue entries
for running jobs when restarting, but "hello" response
only includes an array of job ID's.

Return an array of job objects in response to the hello
request, and in each object include the same data that
would have been received in an alloc request.

Update the hello response parser in libschedutil to
parse the new message format.