shell: add completion ref for each task
Use flux_shell_add_completion_ref() to take a reference on the shell
for each task executed. Drop the reference in the completion callback
so that the shell's reactor is manually stopped after the final task
exits.

This doesn't change any functionality of the shell, but ensures the
shell will still exit the reactor when new reactor references are
attached to the shell in future commits (e.g. message watchers).