shell: use loop:// connector in standalone mode
Problem: leader services don't work in standalone mode.

It's convenient to be able to send messages to ones
self, e.g. so there don't need to be two code paths
for stdio aggregation.  Change standalone mode to use
a valid loop:// broker handle so it is still possible
to send messages to self.