job-exec: avoid SIGKILL of job shells on exceptions
Problem: The job-exec module immediately sends SIGKILL to all
job shells in a job when an exception occurs. This doesn't give
the shells a chance to terminate the tasks of job and perform
other cleanup.

Instead, send SIGTERM (for which the job shell will have a handler)
and then follow up with SIGKILL after a configurable timeout. The
kill timeout is by default 5.0s, but is configurable via a broker
attribute: job-exec.kill_timeout.