shell: subscribe to shell events in initialization
Always subscribe to topic "shell-<id>.*" events and add a new
helper function `flux_shell_add_event_handler` which shell components
can use to add handlers for specific sub-topic events.

The msg_handler_t is not returned to the user and is instead tied
to the flux_t handle for destruction.