testsuite: add lptest program
Add "ripple test" program like lptest(1) from the
lpr package, for testing shell stdio transport.

It's so trivial, it's probably better to write one
and include it as a test program than add a dependency
on the 'lpr' package, complicated by the fact that
cups provides 'lpr' too, but not 'lptest'.