testsuite: use real job shell in attach test
Problem: flux attach will look for output key in KVS
and fail if it doesn't find it, which is incompatible with
the flux-exec test mode that doesn't produce output.
The t2500-job-attach.t combines flux attach and test mode,
and thus will fail.

Alter the test to use the real job shell, which was
unavailable when the test was written.

Drop jq dependency.