schedutil: add a schedutil_t handle
add a schedutil_t struct, which will enable the tracking of outstanding
messages/futures, which, in turn, will enable automatic replies of
ENOSYS on unload and pave the way for simulators to determine when the
schedutil library is idle.

remove ops_context struct and move members into the schedutil_t struct.

refactor all of the public schedutil interfaces to take a schedutil_t
rather than a flux_t as their first argument.

refactor the schedutil library to automatically register services when
the schedutil_t ctx is created. That is, schedutil users no longer need
to call schedutil_ops_register after initializing the library.