flux-shell: populate PMI_process_mapping in PMI KVS
Problem: PMI_process_mapping is unavailable for clique
computation by MPI tasks.

Add logic to generate this key, and add it to the PMI
KVS hash on each shell.

If the key cannot be set because the value exceeds the
static value length, emit a log message on stderr but
don't treat it as an error since the client can run
without clique information.

Fixes #2277