flux-shell: [cleanup] reword PMI inline doc
Problem: this description of callbacks from the PMI protocol
engine was unclear.  It sounds like the whole thing is blocked
while one callback is handled.

Reword to emphasize that only the task that caused the
callback is effectively blocked while awaiting a response.
Other tasks may continue to work in the protocol engine,
and the shell's reactor is never stalled.