flux-shell: don't commit PMI_process_mapping
PMI_process_mapping is added to the local PMI hash
at initialization, and should have the same value
on every shell.  But because it's in the hash, it's
considered "dirty data" to be committed to the shared
KVS during a PMI_Barrier.

Explicitly exclude this key from the KVS transaction
committed during the PMI_Barrier, and thus avoid overhead
that has no effect.