python: refactor Jobspec into a class hierarchy with factory funcs
Recapitalize JobSpec to Jobspec to match the formatting in RFC14.

JobspecV1 is a new class that inherits from Jobspec.

Update the constructors of both to take the top-level keys in the Jobspec
definition.

Add factory functions for creating Jobspec objects from various
combinations of arguments.  This is done since Python does not support
constructor overloading and we do not want to make every argument a
keyword argument.