libflux: buffer: don't invoke read callbacks multiple times
Problem: When buffer data or a line is pending in a flux_buffer_t
the implementation calls the provided read callback in a loop until
no more data is being consumed. While this makes logical sense for
a standalone buffer implementation, this isn't a good fit with how
the buffer is being used within libflux, since in a reactor driven
context we want to be sure to re-enter the reactor for fairness to
other watchers and events.

Therefore, instead of calling the provided callback in a loop, just
invoke the provided callback once. When using a flux_buffer_t
standalone, the user's callback will have to read in a loop until
all data is consumed, or the buffer will have to be checked for
data out of bound. However, there should not be much change for
use of flux_buffer_t within a reactor context.