libsubprocess: support signal of remote process in init state
Problem: When a remote process is still in the INIT state, i.e.
a remote exec request has been sent but the response not yet
received, it is inconvenient for a caller to signal that remote
process, since they must wait until the response is received
that puts that process into the RUNNING state.

Adjust flux_subprocess_kill() to allow a signal to be pending
for remote processes that are not yet RUNNING. The signal will
be forwarded once the state changes to RUNNING, or the signal
will be canceled with error if the process fails to exec or
for any other reason doesn't enter the RUNNING state.