job-info: Fix potential matchtag leak
Problem: Under several circumstances, such as when there is a permission
issue, the job-info module will respond to the original caller with an error,
and then destroy the flux future watching the KVS key.  This could lead to a
matchtag leak, b/c the lookup is never actually canceled.

Solution: In these particular circumstances, send a lookup cancellation before
destroying the future.