libioencode: change rank encoding to string
Problem: translating 'ioencode' objects to RFC 24
data events requires that an integer rank be converted
to string.

Simplify our lives and switch 'ioencode' to use a string
encoding for rank so that it is a valid RFC 24 data event
context object.

For now, retain the interfaces for ioencode/decode that
assume a rank is a single integer.  If we want to "compress"
I/O on a shell before transmitting it to the leader, e.g.
so that one object represents output from multiple tasks,
we can add a new API interface for that.