job-exec: add cancel method for exec implementations
Problem: Only a `kill` method is currently exported by execution
implementation modules. This method could be construed as to either
implement cancelling a running job, or signalling a running job (i.e.
as in the `kill(2)` system call).

Provide an additional `cancel` method which is meant explicitly to
cancel pending work, or shells which have not yet started execution.
To send a signal to a running job, the `kill` method is still provided.

For the bulk_exec module, provide a cancel that removes all pending
commands and releases their resources immediately.