job-exec: rename jobinfo_kill to jobinfo_cancel
Problem: The `jobinfo_kill` function in job-exec module has a confusing
name, since it is meant to terminate or cancel executing jobs, not signal
them. Also, the function sends SIGTERM then SIGKILL to running shells,
but there is a race where shells that are still pending launch aren't
canceled/killed.

Rename the `jobinfo_kill` function to `jobinfo_cancel` to avoid some
naming confusion. Also, if the exec implementation offers a `cancel`
method, use this to cancel any to-be-launched shells before they
are started.

However, if there are commands still pending to be run, the function
doesn't cancel those.