job-exec: fix race with job exception during init
Problem: while a new exec job is "initializating" (creating namespace,
fetching kvs data {R, J}, etc) a job exception could come in and destroy
the initializing job.

Protect the jobinfo structure by taking a reference during initialization.
In the continuation for the initialization composite future, do not continue
with job execution if an exception was recieved.