job-exec: bulk-exec: support mock exceptions
Add the ability to create mock exceptions in the job-exec module
during the initializing phase (before any shells are scheduled to
be launched), or starting phase, when some, but not all, shells
have been launched.

A mock exception is generated if a new jobspec key

 attributes.system.exec.bulkexec.mock_exception

is set to either "init" or "starting".

This allows corner cases for exception handling during normal
job execution to be tested.