job-exec: bulk-exec: don't return empty future in bulk_exec_kill
Problem: bulk_exec_kill() could return an empty composite future
which could never be fulfilled if there are processes on the
exec->processes list, but none are in the INIT or RUNNING state.

Check after looping through all futures to ensure there is at
least one child in the wait_all future and return ENOENT if not.