job-exec: use flux_job_kvs_namespace() to create namespace name
Switch to using libjob's flux_job_kvs_namespace() to create job
guest kvs namespace names, instead of using the dothex notation
of a flux_jobid_t.

This puts job namespace name generation in a single place so that
these names do not have to be generated manually anywhere, as
well as allowing the naming convention to be updated in the future
without breaking any existing code.

As a side effect, this fixes a problematic side effect of kvs
namespace names with periods in them, since these names are
used as keys in kvs module stats output.

Fixes #2306