modules/connector-local: drop response privacy check
Problem: In response_cb(), if there is a send error,
an error is logged only if allowed_message() returns
true, but it always returns true because responses never
set FLUX_MSGFLAG_PRIVATE.  Furthermore, the restriction on
logging is pointless (presumably it was intended to
restrict delivery).

Remove the check and open #2321 on the possible
security vulnerability it leaves open.