broker: set parent-kvs-namespace attribute
Problem: FLUX_KVS_NAMESPACE is cleared in the broker
environment if set, but it is need by rc3.d/01-enclosing-instance.

Eliminate code to explicitly sanitize FLUX_KVS_NAMESPACE and
FLUX_URI in the broker environment, then set the 'sanitize' bit
for FLUX_URI in attrmap, and add FLUX_KVS_NAMESPACE to attrmap
with 'sanitize' set.

Now the attribute 'parent-kvs-namespace' contains
the previous value of FLUX_KVS_NAMESPACE, if it was set.
Rc scripts can use flux getattr to obtain the value.