cmddriver: add --parent option
Problem: It is inconvenient to connect to a parent Flux instance
and access kvs namespace since both FLUX_URI and FLUX_KVS_NAMESPACE
have to be set from the parent-uri and parent-kvs-namespace attributes.

Since this might be a common use case, and confusing for users, add
a `--parent` option to the flux(1) command driver to fetch these
attributes from the current instance on behalf of the user and set
them appropriately in the current environment.

If a connection to the parent instance is successful, also prepend
the parent's configured FLUX_EXEC_PATH and FLUX_MODULE_PATH to the
current env, so that subcommands (besides builtins) will use those
from the parent not the current instance, if they differ.

Allow the --parent option to be used multiple times, to access
granparent, its parent, and so on. Use of --parent from a root instance
does nothing (akin to `cd ..` when you are in the root directory)