cmd/flux-job: Use event watch for job output
In preparation for future live output with job attach, refactor
flux-job to use flux_job_event_watch() for printing output from
a job instead of using a simple lookup of all output.

A small corner case that results in this change is that if the
job never ran (e.g. resource request impossible), this change could
hang as no events will ever occur in the output log.  To resolve this,
we must check to ensure that the job actually started.