shell: add code to manage a plugin stack
Add a simple plugin stack interface for use by the flux-shell.

The plugstack interface supports loading DSOs with arbitrary callback
symbols and an optional plugin name. Callbacks are called in the order
in which plugins are loaded, except in the case of a name collision,
in which case the last plugin loaded will override the original.

This interface alone doesn't support a config file or other method
for configuring how and when plugins are loaded. That is left for a
higher level.