shell: abstract PMI implementation into a plugin
Untangle the PMI implementation from the rest of the shell and turn
it into a shell builtin plugin under the name "pmi".

This allows, in theory, the default PMI implementation to be
overridden by a dynamically loaded plugin.

There is a bit of cheating here since the pmi plugin actually has
access to shell internals, however in order to simplify development
we allow this violation of the shell plugin abstraction for builtins.