shell: rename current io interface to "output"
Problem: The output plugin is contained in a file called "io.c" and
includes a struct and set of functions that use "io" in the name,
when this module really only deals with output. This could cause
confusion when an eventual "input" plugin is added to the shell.

Rename everything that uses "io" in the name to use "output" instead.