libutil/veb: fix vebpred() corner case at size=32
Problem: vebpred(M-1) at M=32 always returns M.

The root cause is left shifting a 32-bit word by the word size,
which is undefined behavior.  Specifically:  on x86_64 with gcc,
in zeros(), ~0<<32 evaluates to ~0 not 0 as expected by
vebpred() -> ones() -> zeros().

Presumably this is not reported by -Wshift-count-overflow
b/c the shift count is computed, not constant.

Special case k==WORD in zeros(), so that the expected result
is returned.

Add assertions in this and a few other areas where the shift
count is computed to ensure it doesn't overflow.

Fixes flux-framework/flux-core#2336

See also: mrdomino/libveb#1