broker: check for optional instance-level PMI key
Check for an optional PMI key "flux.instance-level" which may be
set by the flux-shell of an enclosing instance to tell a child
instance at what "level" it is being started in a hiearachy, i.e.
the number of parent instances up to the root/system instance.

If "flux.instance-level" is found, set this as the "instance-level"
attribute, o/w initialize the instance-level attribute to zero.