librouter/disconnect: add disconnect class
As described in RFC 6, a client may disconnect leaving state held
in various services that it has sent requests to.  Disconnect messages
sent to the services accessed by the client provide the opportunity for
the service to destroy this state or cancel long running operations.
The implementation is based loosely on code from connector-local.

For each outgoing request, disconnect_arm() is called.  If the service
has not been used before, an hash in the disconnect object is "armed"
with a disconnect message for that service.  When the disconnect object
is destroyed, e.g. when the client or its routing table entry are,
the messages are passed one by one to a callback used to send them.