librouter/usock: add class for unix domain sockets
Usock is a new class, based on code in connector-local and
flux-proxy, for managing the service end of a UNIX domain
socket-based Flux connector.  The class assumes that the
Flux reactor is used to handle asynchronous connection
acceptance and piecemeal sending and receiving of messages.

Internally it uses the senfd(), recvfd() functions and struct
iobuf buffers, which in turn depend on flux_msg_encode() and
flux_msg_decode().

The 'struct usock_server' represents a server with a listen socket.
It calls an "acceptor" callback to allow/deny connections and
establish credentials.

The 'struct usock_conn' represents one client connection.
Messages are sent to a client with usock_conn_send(), which
adds them to a queue that is asynchronously transferred to the
wire via reactor callbacks.  Messages are received and assembled
asynchronously by reactor callbacks, then handed to the user via
a callback.

The 'struct usock_conn' has an aux hash which can be used to
attach items that need to be destroyed when the client disconnects,
such as an entry in a routing table.

A special interface allows a connection to be created independent of
the "acceptor", optionally with separate input and output file descriptors.
This is specifically for the flux-relay utility, used by the ssh://
connector which tunnels a local:// connection through ssh standard I/O.

The 'struct usock_client' is a client-side counterpart to the
to the server-side 'struct usock_conn', and is structured to be used
by flux connectors and unit tests.