libflux/msg_handler: don't deliver response with routes
Problem: The message dispatcher does not check whether a
response message has routes remaining in its route stack before
attempting to match an RPC in the matchtag hash.  For a module
like connector-local acting as a router, this could result
in a response being delivered to a connector-local RPC when
it should be routed to a downstream client.

Refuse to match responses to RPC requests (valid matchtag)
unless their route stack is empty.  Other message handlers
for requests continue to match so that routers can be implemented.