librouter/servhash: add servhash class
Add a class that tracks and manages dynamically registered
services for a router.  The router intercepts service.add
and service.remove requests and handles them locally.
A proxy service.add or service.remove is sent to the broker.
While a service is registered, request messages that match
the service are matched to the appropriate client uuid
for forwarding by the router.

This implementation handles the requests asynchronously,
never blocking reactor progress.