shell: Output data as it arrives
Instead of saving all output data and writing it out when a job completes,
write output to the KVS (or stdout/stderr in standalone mode) as output
data arrives to the leader shell.

In this change, every output that is received by the leader shell is
appended to an output array before it is immediately written to the KVS (or
stdout/sterr in standalone mode).  This output array is not explicitly
needed but has been left in tact for future use in aggregating output.