libflux/msg_handler: match request globs last
Problem:  Last-in request handlers have the highest priority,
which allows built-in module methods to be overridden, but prevents
a "fall through" request handler from being registered that would
match methods not explicitly registered earlier.

Instead of pushing all request handlers onto a list, store
non-glob request handlers in a hash by topic string.  Requests
are first matched in the hash, and fall through to the list
if no match.

Registering a handler for a duplicate request method replaces
the old one.  This mimics the current behavior of shadowing
an earlier-registered request handler.