cmd/flux: do not export FLUX_CONF_INTREE
Problem: The FLUX_CONF_INTREE environment variable may be somewhat
brittle as a method to determine if flux components are running in-tree
or not. For example, the variable is exported, but never unset,
and it is unclear what effect this might have on jobs, etc.

Now that there are no users of FLUX_CONF_INTREE, just drop it.