job-info: Fix racy cancellation on transitions
During guest watch state transitions, there is a small racy scenario
in which a user may cancel an event watch while the event watch is
in the process or returning an error (ENODATA, ENOTSUP, etc.).  Detect
this cancellation and respond to the cancellation instead of moving
onto the next state.