flux-broker: drop support for session-id attribute
Problem: the 'session-id' attribute is intended to
be the job ID of the flux instance, but it is passed
through the 32-bit PMI appnum, so it gets mangled.

Drop support for the 'session-id' attribute.

Use the pid of the broker instead of the session id
to construct the rundir and persistdir paths.

Drop session-id from default boot.conf file.

Update config file sharness test and its inputs.