shell: update plugstack implementation for dso loading
Add plugstack_load and plugstack_loadall functions to load a DSO
plugin into the shell's plugin stack or load from a search path
respectively.

Allow a builtin searchpath with getter/setter, and an optional
set of plugstack "aux" items which are propagated to all plugins
loaded via the plugin loading functions of plugstack.