job-info: Workaround namespace remove race
In the kvs-watch module, there is a small possibility the namespace
remove event could be received before the last bits of data from
a watch are sent to a watcher.

- user write to eventlog, kvs commits write, sends setroot event
- user deletes namespace, kvs sends namespace-remove event
- kvs-watch receives setroot event, does lookupat on key
- kvs-watch receives namespace-remove event BEFORE lookupat returns,
  sends ENOTSUP to watchers
- response from lookupat arrives after ENOTSUP sent

To workaround this, in the job-info module, always track how much data
from the guest eventlog has been sent.  When the namespace has been
removed, fallthrough from the guest namespace to the primary KVS
namespace, and check that all data that could be sent to the caller
has been sent.

Fixes #2386