libpmi/pmi2: add minimal PMI-2 API implementation
Problem:  MPICH or derivatives configured --with-pm=slurm
--with-pmi=pmi2 will dlopen Slurm's libpmi2.so in the system
ld.so path even when launched by Flux.

The Slurm library notices PMI_FD set by Flux and attempts
wire protocol version negotation.  It accepts Flux's request
to use protocol version 1, but then proceeds to use protocol
version 2 and becomes sad.

Provide a minimal libpmi2.so implemented on the pmi_simple_client
class that uses protocol version 1.  The strategy will be to
install it outside of the system ld.so path, then set LD_LIBRARY_PATH
in a shell "mpi personality" plugin (once we are ready to have
those) such that Slurm's libpmi2.so with the faulty version
negotiation is bypassed.

Note: going forward, it would be a better solution to support
protocol version 2 (and 1) in the pmi_simple_server class, but
that is a non-trivial undertaking not attempted here.