shell: Add stdout/stderr output type config
In preparation for supporting multiple output types in the shell,
support an internal configuration for stdout/stderr output types.  The
output types are either "term" (for writing to terminal) or "kvs".
There is no behavior change with the current implementation.

Adjust the output builtin to not assume it is always handling stdout
and stderr.  Only perform setup/config if data will be going to the
terminal or KVS.  Adjust functions to handle the possibility that
they may handle only one of stdout/stderr, but not both.

Adjust logic to call shell_output_kvs_init() as long as we're not
standalone.