shell: Support stdout/stderr file output
Support stdout/stderr file redirection in jobs via the

attributes.system.shell.options.output.<stream>.type

jobspec setting and the type "file".  This option will redirect all
stdout/stderr from all tasks into a single file.  A path to the file can be
specified via:

attributes.system.shell.options.output.<stream>.path

As a custom, when a stdout type is specified, but a stderr type is not, the
stderr type will follow the stdout type.  This includes writing to the
same file.

Note that this initial implementation blocks on calls to write(2).