shell/pmi: use jobid as kvsname
Problem: kvsname is hardwired to the string "pmi".  Other PMI
implementations use the jobid or jobid.<step>.  Furthermore,
PMI-2 names this field "jobid".  Since we just added a 'flux.jobid'
local PMI key, perhaps we should conform and use this instead,
then avoid the extra request/response during Flux-in-flux bootstrap.

Use a string representation of the Flux jobid for the kvsname
when PMI is offered by the Flux shell.