shell: Refactor output initialization
In the shell output builtin, split up the primary shell_output_flush()
and shell_output_commit() functions into "init" and "core" functions.
The "init" functions handle the output header and/or any eventlog entries
that need to be written out.  The "core" functions handle the actual
processing of data.

As a result, the output header does not need to be written to the output
array.  Other flag variables such as "output_ready_sent" are
no longer needed as well.