shell: Add output type to eventlog
In addition to the "output-ready" event in the exec.eventlog, also
add eventlog entries to indicate what types of output are being
done for stdout and stderr via the "output-stdout" and "output-stderr"
events.  Both will include a "type" field to indicate the way in which
stdout/stderr will be output.

For the time being, the only type is "kvs", and all output defaults to
this.

For consistency, the "output-ready" key is now the "output-kvs-ready" and
the shell_output_ready() function is now shell_output_eventlog().