flux-job: Update attach for live output
Adjust attach for fact that output is now live and occurs as the
job is happening.  Watch exec.eventlog to determine when the guest.output
directory is ready for reading.  Once it is ready, watch the guest.output
and output as data is available, and not after the job completes.