cmd/flux-job: add attach --show-exec
Add an attach option to display eventlog entries from the
guest.exec.eventlog, in the same format as the main
job eventlog entries.

Change the logic slightly to:
* Request the guest eventlog AFTER the main eventlog 'submit'
  event is received.  This ensures guest events do not arrive
  before the timestamp zero point can be established by submit.

* Don't cancel the guest eventlog after the 'output-ready' event.
  Continue processing guest events for display.

Note that if both --show-exec and --show-events are specified,
events from the two logs may be displayed out of time order,
especially if the logs are not being produced in real time as
they are consumed.