shell: add builtin cpu affinity
Add a builtin 'affinity' plugin for the shell which sets affinity
for the job as a whole to the allocated list of cores by default,
using hwloc to find the cpuset that includes all assigned cores
and their associated PUs (if any).

The plugin looks for the shell option 'cpu-affinity' and if set to
"per-task", then hwloc_distrib(3) is used to distribute tasks
across the assigned reosurces.

If the cpu-affinity shell option is set to "off" then the plugin is
disabled.

If the cores list presented to the affinity plugin does not intersect
with the real cores available on the machine, then the affinity
plugin logs an informational message and gives up.

Fixes #2243