testsuite: move long running tests first in TESTS
Move the long running tests t5000-valgrind.t and t3100-flux-in-flux.t
to the beginning of the TESTS list, so that they are started first
in `make check`. This should result in a lower makespan for `make check`
for `-j` > 1. (Especially on Travis where we can only run with -j2).