libflux/conf: prepend top_builddir/t to in-tree LUA_PATH
Problem: on some Lua 5.2 systems, the current directory is
last in the default LUA_PATH instead of first. This could result in
flux Lua tests picking up the installed version of the test
driver `fluxometer.lua` instead of the in-tree version.

For in-tree flux(1), add the *explicit* path to fluxometer.lua,
"$(top_builddir)/t/?.lua", to the "lua_path_add" libflux conf
setting, to better ensure the right fluxometer is found in
tests.

Fixes #2403