testsuite: always prepend builddir/t to LUA_PATH
Problem: on some systems `./?.lua` (the currend dir) is not always
first in the default LUA_PATH. The lua-based tests in the testsuite
depend on having this directory first in the path to ensure the
correct fluxometer.lua is found during `make check`.

Instead of assuming that the default empty LUA_PATH will work
correctly, _always_ prepend "$(buiddir)/t/?.lua" to LUA_PATH in the
testuite Makefile.am, to ensure that an incorrect system installed
fluxometer.lua is not found during testing.