bindings/python: allow CLIMain log level to be set
Problem: sometimes errors are hard to track down with
the CLIMain default exception handler, which only prints
one line with no indication of what part of the code is
failing.

Since stack traces are still logged at DEBUG level, and
the default log level in CLIMain is INFO, add support for
an optional environment variable which can alter the debug
level.

Set FLUX_PYCLI_LOGLEVEL=10 to change the logging level to DEBUG.