shell: add builtin gpubind plugin for the shell
Add a builtin gpu affinity plugin for the shell which sets
CUDA_VISIBLE_DEVICES (optionally per-task), as well as
CUDA_DEVICE_ORDER=PCI_BUS_ID.

This builtin plugin can be overridden with the name 'gpu-affinity'.

The plugin reads the shell option "gpu-affinity: string", and supports
values "on", "off", and "per-task". Invalid options are ignored
and the default is "on".