modules/job-manager: add event_job_post_pack() NOLOG flag
Problem: event_job_post_pack() code is duplicated in submit.c
for the submit event, special-cased because the eventlog entry
was already logged by the job-ingest module.

Add a flag to event_job_post_pack() to skip the eventlog part
of the update, and call it with the flag from submit.

The event_batch_pub_state() function is no longer needed outside
of event.c, so change to static scope and remove from event.h.