modules/job-manager: retain inactive jobs in the queue
Problem: it is inconvenient to have no mechanism to list
inactive (completed) jobs.

Instead of removing jobs from the queue when they transition
to inactive, leave them there so they can be listed by
flux job list.

Add checks to priority and raise handlers to reject the
request if the target job is inactive.