sched-simple: libjj: catch all unsupported resource types
Problem: When parsing jobspec with libjj, all but the first resource
vertex in a list under an out edge was read by the parser, and other
resources were ignored. Thus, requests for resources of unsupported
types were being accepted by the scheduler, and the unsupported types
ignored (as long as they were not the first element in the resource
list)

Change the libjj implementation to read all resources in the list at
each level, and return an error if any resource doesn't match the
list of supported types.

Fixes #2417.