test: t0001-basic: fix incorrect flux-start ARGS
In 91b493bf13940020526ca3af19fb6f071528bbbc the update to ARGS
to add the rc1,rc3 paths overwrote ARGS instead of appending to
it, causing other desired flux-start ARGS to be dropped.
(Also there was a typo in `rc3_path`, but we'll pretend that
never happened)

Append to ARGS instead of overwriting it in the tests.