test: t1003-kvs-stress.t: storing 1M keys in a dir takes too long
Problem: The test 'kvs: store 1,000,000 keys in one dir' takes
an extremely long time to complete. Even though this test is marked
with the LONGTEST prereq, it is too long to be appropriate for
a 'make check' based testsuite.

Reduce the number of keys by an order of magnitude to get a
resonably long test.

Fixes #2442