testsuite: extend too small run_timeout test timeouts
Problem: On a slow system, the tests for "run_timeout" may fail
when due to the 10ms timer firing before the actual test
executable (or lack of valid executable) is able to be run.

Extend the timeout to 1s -- this timeout should only be hit if the
test fails anyway.