flux-job: Refactor attach eventlog watch check
Instead of having two counters for eventlog watchers that have
started and completed, just have a single counter.  The job will
be considered complete once the counter has reached zero.

In addition, change the name of the attach_eventlog_watch_completed_check()
function to attach_completed_check() so it is more generic.