connectors/local: include socket path in URI
Problem: the convention of having the local:// connector
URI refer to a directory containing a socket named "local"
is inconvenient when we try to use ssh socket forwarding
for the local connector.

Switch the URI to refer to the socket path rather than
the directory containing it.

Update the broker, flux-proxy, and connector-local
to use this convention also.

Fixes #2299.