cmd/flux-remote: new command
Add a util that launches a remote command with ssh, forwarding
the local broker socket to the remote host, and running the
command under flux-proxy so that the remote command runs as if
it were spawned within the local environment:  in the same working
directory, using the same FLUX_KVS_NAMESPACE (if any), and with
FLUX_URI so so that it ultimately leads back to the local broker.

This command relies on the Unix domain socket forwarding
capability added to OpenSSH in version 6.7 (relased 2014-ish).
It requires that the server be configured to allow that.

This is one step towards enabling the job-exec system to launch
remote flux shells on systems that don't run a broker on every
compute node.