cmd/flux-proxy: add --chdir, --setenv  options
Problem: when running flux-proxy remotely via ssh, it's
inconvenient to manipulate the environment and current
working directory using only ssh.

To set environment variables in ssh, one must whitelist the
variables in the sshd config using AcceptEnv.

There is no mechanism to set the current working directory
is the ssh remote command, other than inserting a "cd" command
in front of it.