modules/sched-simple: fetch resources without WAITCREATE
Problem: if sched-simple load loaded with a res=KEY argument
that doesn't exist, it hangs forever.

Resources are fetched from the KVS wiht the FLUX_KVS_WAITCREATE
flag so that aggregation of hwloc data could run in parallel
with other rc1 activities, and sched-simple, its consumer, would
block until it was ready.

Now that arbitrary resource keys can be specified on the command
line, this is not great behavior.  Drop the flag, and change rc1
so that the shell waits for the hwloc operation to complete just
before loading sched-simple.