modules/job-exec: add targets=KEY option
Add an option to the job-exec module that lets some
execution targets (ranks) be mapped to hostnames.
The option points to a KVS key whose value is a dictionary
of rank => hostname.

If an execution target is not in the targets dictionary,
or if the targets option was not specified, then presume
that the target is a broker rank, as before, and use the
broker exec service to directly launch the flux-shell.

Otherwise use flux-remote to launch the flux-shell on
a remote host.