shell: ensure shell service name is unique
Problem: each shell tries to register a service name of
"shell-<jobid>", but that is no longer unique when there
are multiple shells launched per broker.

Append the shell rank to the service name on follower shells.
Leave the leader shell alone, as front end commands
like flux-job attach need to communicate with it.