shell: assume broker 0 ran all remote shells
In order for shell peer-to-peer services to work properly,
we need to be able to map the shell rank to the broker rank
that the shell was spawned from.

'rcalc' can translate shell rank to exec target rank.

We have no reliable way in the shell to translate exec target
rank to broker rank.  Right now only job-exec has this mapping.
While we puzzle out how to provide this mapping to shells,
implement the following hack:

If the instance size is 1, assume the broker rank is 0 for
all execution targets.  This is true both for native and foreign
resouces.  If the instance size is greater than 1, assume the
broker rank is equal to the exec target rank.  This is true only
for native resources.

Therefore, for the time being, only an instance size=1 can launch
on foreign resources.