libutil: intree: fix crash when builddir can't be accessed
Problem: A bug in is_intree() causes NULL builddir to be passed
to strlen when realpath(3) returns ENOENT or EACCESS, instead of
immediately returning 0 or false.

Improve the code to immediately return 0 on EACCES or ENOENT errors
from realpath and -1 for any other error from realpath(3) which
causes a NULL return from the function.

Fixes #2467